+++ src/LSST/M1M3/SS/BusLists/ActiveBusList.cpp
@@ -48,7 +48,7 @@ void ActiveBusList::buildBuffer() {
     BusList::buildBuffer();
     SPDLOG_DEBUG("ActiveBusList: buildBuffer()");
 
-    bool slewFlag = BoosterValveStatus::instance().slewFlag;
+    uint8_t boosterValves = BoosterValveStatus::instance().opened ? 255 : 0;
 
     _lvdtSampleClock = 0;
     for (int subnetIndex = 0; subnetIndex < SUBNET_COUNT; subnetIndex++) {
@@ -83,7 +83,8 @@ void ActiveBusList::buildBuffer() {
                 }
             }
             this->ilcMessageFactory->broadcastForceDemand(&this->buffer, _outerLoopData->broadcastCounter,
-                                                          slewFlag, saaPrimary, daaPrimary, daaSecondary);
+                                                          boosterValves, saaPrimary, daaPrimary,
+                                                          daaSecondary);
             this->buffer.writeTimestamp();
             for (int faIndex = 0; faIndex < this->subnetData->getFACount(subnetIndex); faIndex++) {
                 uint8_t address = this->subnetData->getFAIndex(subnetIndex, faIndex).Address;
@@ -151,7 +152,7 @@ void ActiveBusList::buildBuffer() {
 void ActiveBusList::update() {
     _outerLoopData->broadcastCounter = RoundRobin::BroadcastCounter(_outerLoopData->broadcastCounter);
 
-    bool slewFlag = BoosterValveStatus::instance().slewFlag;
+    uint8_t boosterValves = BoosterValveStatus::instance().opened ? 255 : 0;
 
     for (int subnetIndex = 0; subnetIndex < SUBNET_COUNT; subnetIndex++) {
         if (this->subnetData->getFACount(subnetIndex) > 0) {
@@ -178,7 +179,8 @@ void ActiveBusList::update() {
             }
             this->buffer.setIndex(_setForceCommandIndex[subnetIndex]);
             this->ilcMessageFactory->broadcastForceDemand(&this->buffer, _outerLoopData->broadcastCounter,
-                                                          slewFlag, saaPrimary, daaPrimary, daaSecondary);
+                                                          boosterValves, saaPrimary, daaPrimary,
+                                                          daaSecondary);
 
             int32_t statusIndex = _roundRobinFAReportServerStatusIndex[subnetIndex];
             int32_t dataIndex = this->subnetData->getFAIndex(subnetIndex, statusIndex).DataIndex;