+++ src/LSST/M1M3/SS/Commands/ClearOffsetForcesCommand.cpp
@@ -29,8 +29,7 @@ namespace LSST {
 namespace M1M3 {
 namespace SS {
 
-ClearOffsetForcesCommand::ClearOffsetForcesCommand(int32_t commandID, MTM1M3_command_clearOffsetForcesC*)
-        : Command(commandID) {}
+ClearOffsetForcesCommand::ClearOffsetForcesCommand(int32_t commandID) : Command(commandID) {}
 
 void ClearOffsetForcesCommand::execute() { Context::get().clearOffsetForces(this); }